+++ Mindbox/Model/Response/ProductListResponse.swift
@@ -3,6 +3,6 @@ import Foundation
 open class ProductListResponse: Decodable {
     public let product: ProductResponse?
     public let productGroup: ProductGroupResponse?
-    public let count: Decimal?
-    public let priceOfLine: Decimal?
+    public let count: Double?
+    public let priceOfLine: Double?
 }
